Reduce time needed to save a measurement
When a page uses nearly 100% CPU saving measurement data
is challenging. To maximize the chance to save data, the priority
of mongoose thread have been increase.
The BHO has also been optimize to stop looking for page
load over when the page has been loaded.
Finally, saving images is now time bound. The hook as 10s
to save all images for a given step. On a normal CPU usage
this represents ~ 100 different images.