Fix resources bleeding into next step
To avoid bleeding of resources from one step to the next, all requests
between the first step request and its response should be pushed back to
the previous step (blackout period).
The rational being, if the `index.html` is currently requested, none of
the other request can possibly be part of this new step since it's not
even loaded.